Break out common stuff into a new python library
